+++ packages/components/src/DataList/components/DataListLayoutInternal/DataListHeader.tsx
@@ -11,32 +11,36 @@ import {
 import { sortSizeProp } from "../../DataList.utils";
 import { DataListLayoutProps } from "../DataListLayout/DataListLayout";
 
+interface DataListHeaderProps<T extends DataListObject> {
+  layouts: React.ReactElement<DataListLayoutProps<T>>[] | undefined;
+  mediaMatches?: Record<Breakpoints, boolean>;
+  headerData?: DataListItemTypeFromHeader<T, DataListHeaderType<T>>;
+  headerVisibility: NonNullable<DataListProps<T>["headerVisibility"]>;
+}
+
 export function DataListHeader<T extends DataListObject>({
   layouts,
   mediaMatches,
   headerData,
   headerVisibility,
-}: {
-  layouts: React.ReactElement<DataListLayoutProps<T>>[] | undefined;
-  mediaMatches?: Record<Breakpoints, boolean>;
-  headerData?: DataListItemTypeFromHeader<T, DataListHeaderType<T>>;
-  headerVisibility: NonNullable<DataListProps<T>["headerVisibility"]>;
-}) {
+}: DataListHeaderProps<T>) {
+  const matchingMediaQueries = Object.keys(mediaMatches || {}).filter(
+    (key): key is Breakpoints => !!mediaMatches?.[key as Breakpoints],
+  );
+  const sortedVisibleBreakpoints = sortSizeProp(matchingMediaQueries);
+
+  // Determines if the header should be visible based on the headerVisibility
+  const showHeader = sortedVisibleBreakpoints.reduce(
+    (previousVisibility, breakpoint) => {
+      return headerVisibility[breakpoint] ?? previousVisibility;
+    },
+    true,
+  );
+
   return (
     <DataListLayoutInternal
       layouts={layouts}
-      renderLayout={(layout: React.ReactElement<DataListLayoutProps<T>>) => {
-        const matchingMediaQueries = Object.keys(mediaMatches || {}).filter(
-          (key): key is Breakpoints => !!mediaMatches?.[key as Breakpoints],
-        );
-        const sortedVisibleBreakpoints = sortSizeProp(matchingMediaQueries);
-        // Determines if the header should be visible based on the headerVisibility
-        const showHeader = sortedVisibleBreakpoints.reduce(
-          (previousVisibility, breakpoint) => {
-            return headerVisibility[breakpoint] ?? previousVisibility;
-          },
-          true,
-        );
+      renderLayout={layout => {
         return (
           (showHeader && headerData && (
             <div className={styles.header}>
