+++ packages/components/src/DataList/components/DataListLayoutInternal/DataListLayoutInternal.tsx
@@ -4,6 +4,38 @@ import { DataListObject } from "../../DataList.types";
 import { sortSizeProp } from "../../DataList.utils";
 import { DataListLayoutProps } from "../DataListLayout/DataListLayout";
 
+interface DataListLayoutInternalProps<T extends DataListObject> {
+  layouts: React.ReactElement<DataListLayoutProps<T>>[] | undefined;
+  renderLayout: (
+    layout: React.ReactElement<DataListLayoutProps<T>>,
+  ) => JSX.Element;
+  mediaMatches?: Record<Breakpoints, boolean>;
+}
+
+export function DataListLayoutInternal<T extends DataListObject>({
+  layouts,
+  renderLayout,
+  mediaMatches,
+}: DataListLayoutInternalProps<T>) {
+  const sizePropsOfLayouts = layouts?.map(layout => layout.props.size || "xs");
+  const layoutToRender = layouts?.find(layout => {
+    const layoutSize = layout.props.size || "xs";
+    const isVisible = isLayoutVisible({
+      layoutSize,
+      mediaMatches,
+      sizePropsOfLayouts,
+    });
+    return isVisible;
+  });
+  return (layoutToRender && renderLayout(layoutToRender)) || <></>;
+}
+
+interface IsLayoutVisibleParams {
+  layoutSize: Breakpoints;
+  mediaMatches?: Record<Breakpoints, boolean>;
+  sizePropsOfLayouts?: Breakpoints[];
+}
+
 /**
  * Determine is layout is visible by checking a media query matches for the
  * visible sizes of the layout and there isn't a larger layout that should be rendered
@@ -12,11 +44,7 @@ function isLayoutVisible({
   layoutSize,
   mediaMatches,
   sizePropsOfLayouts,
-}: {
-  layoutSize: Breakpoints;
-  mediaMatches?: Record<Breakpoints, boolean>;
-  sizePropsOfLayouts?: Breakpoints[];
-}) {
+}: IsLayoutVisibleParams) {
   const largerBreakpointsToHide = sizePropsOfLayouts?.filter(
     size => BREAKPOINTS.indexOf(size) > BREAKPOINTS.indexOf(layoutSize),
   );
@@ -41,28 +69,4 @@ function isLayoutVisible({
   );
 
   return isVisibleBreakpoint && largerLayoutIsNotVisible;
-}
-
-export function DataListLayoutInternal<T extends DataListObject>({
-  layouts,
-  renderLayout,
-  mediaMatches,
-}: {
-  layouts: React.ReactElement<DataListLayoutProps<T>>[] | undefined;
-  renderLayout: (
-    layout: React.ReactElement<DataListLayoutProps<T>>,
-  ) => JSX.Element;
-  mediaMatches?: Record<Breakpoints, boolean>;
-}) {
-  const sizePropsOfLayouts = layouts?.map(layout => layout.props.size || "xs");
-  const layoutToRender = layouts?.find(layout => {
-    const layoutSize = layout.props.size || "xs";
-    const isVisible = isLayoutVisible({
-      layoutSize,
-      mediaMatches,
-      sizePropsOfLayouts,
-    });
-    return isVisible;
-  });
-  return (layoutToRender && renderLayout(layoutToRender)) || <></>;
 }
